+++ src/unimport/__init__.py
@@ -1,2 +1,2 @@
-__version__ = "0.13.0"
+__version__ = "0.14.0"
 __description__ = "A linter, formatter for finding and removing unused import statements."